requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Azure OpenAI
openai==1.12.0

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator>=2.0.0,!=2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
